+++ src/libdessert/dessert_ping.c
@@ -95,7 +95,7 @@ int dessert_cli_cmd_ping(struct cli_def* cli, char* command, char* argv[], int a
  * @retval DESSERT_MSG_DROP if the ping is destined to this host
  * @retval DESSERT_MSG_KEEP if some other host is the destination
  */
-dessert_cb_result dessert_mesh_ping(dessert_msg_t* msg, size_t len, dessert_msg_proc_t* proc, dessert_meshif_t* meshif, dessert_frameid_t id) {
+dessert_cb_result dessert_mesh_ping(dessert_msg_t* msg, uint32_t len, dessert_msg_proc_t* proc, dessert_meshif_t* meshif, dessert_frameid_t id) {
     dessert_ext_t* ext;
     struct ether_header* l25h;
     u_char temp[ETHER_ADDR_LEN];
@@ -132,7 +132,7 @@ dessert_cb_result dessert_mesh_ping(dessert_msg_t* msg, size_t len, dessert_msg_
  * @retval DESSERT_MSG_DROP if the pong is destined to this host
  * @retval DESSERT_MSG_KEEP if some other host is the destination
  */
-dessert_cb_result dessert_mesh_pong(dessert_msg_t* msg, size_t len, dessert_msg_proc_t* proc, dessert_meshif_t* meshif, dessert_frameid_t id) {
+dessert_cb_result dessert_mesh_pong(dessert_msg_t* msg, uint32_t len, dessert_msg_proc_t* proc, dessert_meshif_t* meshif, dessert_frameid_t id) {
     dessert_ext_t* ext;
     struct ether_header* l25h;
     u_char temp[ETHER_ADDR_LEN];
